feat: add implementation for `Status`
- Implement `Status` feature
- Found bug in which the stream is blocking

The blocking issue occurs on the `stream.peek()` method in the
`stream_into_vec` function. What I have discovered are these:
- The data would be perfectly processed after the block
- It blocks after the server sent a response for the status request

The only possibility that I can think of is an issue with the stream
being split into two parts. I haven't tried it yet since I wanted to
commit this first before breaking everything, but it'd be a funny one
if this message aged well.