feat: make `read_from` use a generic reader
the main reason to do this is so that we have a function that's also
able to read from the `TcpStream`, thus making it so that the length is
consumed and there's no unnecessary peeking. you can still check the
old code in this commit as i haven't changed it (because i can't test
anything) and see that the length is basically unneeded and constructed
twice for basically nothing.