feat: play handle (and many others)
if you're for some reason is reading through the commits, i apologize
for this one as i couldn't focus on one thing to do and kept on
switching task which resulted in this huge pile of changes. to properly
focus on the task on hand shall now be one of the objective of this
project now.